+++ packages/nimble-components/src/wafer-map/modules/rendering.ts
@@ -10,11 +10,24 @@ export class RenderingModule {
 
     public constructor(private readonly wafermap: WaferMap) {}
 
+    public updateSortedDiesAndDrawWafer(): void {
+        this.dies = this.wafermap.dataManager.diesRenderInfo.sort((a, b) => {
+            if (a.fillStyle > b.fillStyle) {
+                return 1;
+            }
+            if (b.fillStyle > a.fillStyle) {
+                return -1;
+            }
+
+            return 0;
+        });
+        this.drawWafer();
+    }
+
     public drawWafer(): void {
         this.wafermap.canvasContext.save();
         this.clearCanvas();
         this.scaleCanvas();
-        this.sortDies();
         this.renderDies();
         this.renderText();
         this.wafermap.canvasContext.restore();
@@ -49,61 +62,75 @@ export class RenderingModule {
         return '';
     }
 
-    private sortDies(): void {
-        this.dies = this.wafermap.dataManager.diesRenderInfo.sort((a, b) => {
-            if (a.fillStyle > b.fillStyle) {
-                return 1;
-            }
-            if (b.fillStyle > a.fillStyle) {
-                return -1;
-            }
-
-            return 0;
-        });
-    }
-
     private renderDies(): void {
-        let prev: DieRenderInfo | undefined;
+        const context = this.wafermap.canvasContext;
+        const dieWidth = this.wafermap.dataManager.dieDimensions.width;
+        const dieHeight = this.wafermap.dataManager.dieDimensions.height;
+        const transformedCanvasMinPoint = this.wafermap.transform.invert([
+            0, 0
+        ]);
+        const transformedCanvasMaxPoint = this.wafermap.transform.invert([
+            this.wafermap.canvas.width,
+            this.wafermap.canvas.height
+        ]);
+        transformedCanvasMinPoint[0] -= dieWidth;
+        transformedCanvasMinPoint[1] -= dieHeight;
 
         for (const die of this.dies) {
-            if (!prev) {
-                this.wafermap.canvasContext.fillStyle = die.fillStyle;
+            if (
+                this.isDieVisible(
+                    die,
+                    transformedCanvasMinPoint,
+                    transformedCanvasMaxPoint
+                )
+            ) {
+                context.fillStyle = die.fillStyle;
+                context.fillRect(die.x, die.y, dieWidth, dieHeight);
             }
-            if (prev && die.fillStyle !== prev.fillStyle && die.fillStyle) {
-                this.wafermap.canvasContext.fillStyle = die.fillStyle;
-            }
-            this.wafermap.canvasContext.fillRect(
-                die.x,
-                die.y,
-                this.wafermap.dataManager.dieDimensions.width,
-                this.wafermap.dataManager.dieDimensions.height
-            );
-            prev = die;
         }
     }
 
     private renderText(): void {
-        const dieSize = this.wafermap.dataManager.dieDimensions.width
-            * this.wafermap.dataManager.dieDimensions.height
-            * (this.wafermap.transform.k || 1);
-        const fontsize = this.wafermap.dataManager.labelsFontSize;
-        this.wafermap.canvasContext.font = `${fontsize.toString()}px sans-serif`;
-        this.wafermap.canvasContext.fillStyle = '#ffffff';
-        this.wafermap.canvasContext.textAlign = 'center';
-        this.wafermap.canvasContext.lineCap = 'butt';
-        const approxTextHeight = this.wafermap.canvasContext.measureText('M');
-
-        const dieDimensions = this.wafermap.dataManager.dieDimensions;
+        if (this.wafermap.dieLabelsHidden) {
+            return;
+        }
+        const dieWidth = this.wafermap.dataManager.dieDimensions.width;
+        const dieHeight = this.wafermap.dataManager.dieDimensions.height;
+        const dieSize = dieWidth * dieHeight * (this.wafermap.transform.k || 1);
         if (dieSize >= this.minDieDim) {
+            const fontsize = this.wafermap.dataManager.labelsFontSize;
+            const context = this.wafermap.canvasContext;
+            context.font = `${fontsize.toString()}px sans-serif`;
+            context.fillStyle = '#ffffff';
+            context.textAlign = 'center';
+            context.lineCap = 'butt';
+            const approximateTextHeight = context.measureText('M');
+
+            const transformedCanvasMinPoint = this.wafermap.transform.invert([
+                0, 0
+            ]);
+            const transformedCanvasMaxPoint = this.wafermap.transform.invert([
+                this.wafermap.canvas.width,
+                this.wafermap.canvas.height
+            ]);
+            transformedCanvasMinPoint[0] -= dieWidth;
+            transformedCanvasMinPoint[1] -= dieHeight;
+
             for (const die of this.dies) {
-                this.wafermap.canvasContext.fillText(
-                    die.text,
-                    die.x + dieDimensions.width / 2,
-                    die.y
-                        + dieDimensions.height / 2
-                        + approxTextHeight.width / 2,
-                    dieDimensions.width - (dieDimensions.width / 100) * 20
-                );
+                if (
+                    this.isDieVisible(
+                        die,
+                        transformedCanvasMinPoint,
+                        transformedCanvasMaxPoint
+                    )
+                ) {
+                    context.fillText(
+                        die.text,
+                        die.x + dieWidth / 2,
+                        die.y + dieHeight / 2 + approximateTextHeight.width / 2,
+                        dieWidth - (dieWidth / 100) * 20
+                    );
+                }
             }
         }
     }
@@ -112,8 +139,8 @@ export class RenderingModule {
         this.wafermap.canvasContext.clearRect(
             0,
             0,
-            this.wafermap.canvasWidth * this.wafermap.transform.k,
-            this.wafermap.canvasHeight * this.wafermap.transform.k
+            this.wafermap.canvas.width,
+            this.wafermap.canvas.height
         );
     }
 
@@ -125,6 +152,19 @@ export class RenderingModule {
         this.wafermap.canvasContext.scale(
             this.wafermap.transform.k,
             this.wafermap.transform.k
+        );
+    }
+
+    private isDieVisible(
+        die: DieRenderInfo,
+        minPoint: [number, number],
+        maxPoint: [number, number]
+    ): boolean {
+        return (
+            die.x >= minPoint[0]
+            && die.x < maxPoint[0]
+            && die.y >= minPoint[1]
+            && die.y < maxPoint[1]
         );
     }
 }